Initial Commit
This commit provides an HTML page with a div box. The main.css handles
the key down event and moves the box by 10px to the direction of
whichever arrow key that was pressed.

The movement of the div box toward the right side of the viewport is
restricted.

The karma.conf.js file contains the config required to start karma
testing suite on the project.

The tests provided currently tests whether the DOM element is moving
in each direction, and that the movement along right direction is
restricted at the boundary.